vietcardocr/requirements.txt
vietocr==0.3.5
numpy==1.19.5
Unidecode==1.1.1
torch==1.9.0
torchvision==0.10.0
opencv_python==4.5.2.52
Pillow==8.3.1
